fix: make sync naming consistent
