+++ example/client.py
@@ -1,100 +1,55 @@
-from connecpy.exceptions import ConnecpyException, ConnecpyServerException
+import asyncio
+
+import httpx
+
+from connecpy.exceptions import ConnecpyServerException
 
 import haberdasher_connecpy
 import haberdasher_pb2
 
 
 server_url = "http://localhost:3000"
-timeout_ms = 5000
+timeout_s = 5
 
 
-def create_large_request():
-    """Create a request with a large description to test compression."""
-    return haberdasher_pb2.Size(
-        inches=12,
-        description="A" * 2048,  # Add a 2KB string to ensure compression is worthwhile
-    )
+async def main():
+    async with httpx.AsyncClient(
+        base_url=server_url,
+        timeout=timeout_s,
+    ) as session:
+        # Example 1: POST request with Zstandard compression, receiving Brotli compressed response
+        async with haberdasher_connecpy.HaberdasherClient(
+            server_url,
+            session=session,
+            send_compression="zstd",
+            accept_compression=("br",),
+        ) as client:
+            try:
+                response = await client.MakeHat(
+                    request=haberdasher_pb2.Size(inches=12),
+                    headers={
+                        "Content-Encoding": "zstd",  # Request compression
+                    },
+                )
+                print("POST with Zstandard and Brotli compression:", response)
+            except ConnecpyServerException as e:
+                print(e.code, e.message, e.to_dict())
 
-
-def main():
-    # Example 1: POST request with gzip compression (large request)
-    with haberdasher_connecpy.HaberdasherClient(
-        server_url,
-        timeout_ms=timeout_ms,
-        send_compression="gzip",
-        accept_compression=("gzip",),
-    ) as client:
-        try:
-            print("\nTesting POST request with gzip compression...")
-            response = client.MakeHat(
-                request=create_large_request(),
-            )
-            print("POST with gzip compression successful:", response)
-        except (ConnecpyException, ConnecpyServerException) as e:
-            print("POST with gzip compression failed:", str(e))
-
-    # Example 2: POST request with brotli compression (large request)
-    with haberdasher_connecpy.HaberdasherClient(
-        server_url,
-        timeout_ms=timeout_ms,
-        send_compression="br",
-        accept_compression=("br",),
-    ) as client:
-        try:
-            print("\nTesting POST request with brotli compression...")
-            response = client.MakeHat(
-                request=create_large_request(),
-            )
-            print("POST with brotli compression successful:", response)
-        except (ConnecpyException, ConnecpyServerException) as e:
-            print("POST with brotli compression failed:", str(e))
-
-    # Example 3: GET request without compression
-    with haberdasher_connecpy.HaberdasherClient(
-        server_url, timeout_ms=timeout_ms, accept_compression=()
-    ) as client:
-        try:
-            print("\nTesting GET request without compression...")
-            response = client.MakeHat(
-                request=haberdasher_pb2.Size(inches=8),  # Small request
-                use_get=True,
-            )
-            print("GET without compression successful:", response)
-        except (ConnecpyException, ConnecpyServerException) as e:
-            print("GET without compression failed:", str(e))
-
-    # Example 4: GET request with ztstd compression (large request)
-    with haberdasher_connecpy.HaberdasherClient(
-        server_url,
-        timeout_ms=timeout_ms,
-        send_compression="zstd",
-        accept_compression=("zstd",),
-    ) as client:
-        try:
-            print("\nTesting GET request with gzip compression...")
-            response = client.MakeHat(
-                request=create_large_request(),
-                use_get=True,
-            )
-            print("GET with zstd compression successful:", response)
-        except (ConnecpyException, ConnecpyServerException) as e:
-            print("GET with zstd compression failed:", str(e))
-
-    # Example 5: Test multiple accepted encodings
-    with haberdasher_connecpy.HaberdasherClient(
-        server_url,
-        timeout_ms=timeout_ms,
-        send_compression="br",
-    ) as client:
-        try:
-            print("\nTesting POST with multiple accepted encodings...")
-            response = client.MakeHat(
-                request=create_large_request(),
-            )
-            print("POST with multiple encodings successful:", response)
-        except (ConnecpyException, ConnecpyServerException) as e:
-            print("POST with multiple encodings failed:", str(e))
+        # Example 2: GET request, receiving Zstandard compressed response
+        async with haberdasher_connecpy.HaberdasherClient(
+            server_url,
+            session=session,
+            accept_compression=["zstd"],
+        ) as client:
+            try:
+                response = await client.MakeHat(
+                    request=haberdasher_pb2.Size(inches=8),
+                    use_get=True,  # Enable GET request
+                )
+                print("\nGET with Zstandard compression:", response)
+            except ConnecpyServerException as e:
+                print(e.code, e.message, e.to_dict())
 
 
 if __name__ == "__main__":
-    main()
+    asyncio.run(main())
